% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_hull_clustering3D.R
\name{alpha_hull_clustering3D}
\alias{alpha_hull_clustering3D}
\title{Find cell clusters in 3D spatial data using alpha hull clustering
algorithm.}
\usage{
alpha_hull_clustering3D(
  spe,
  cell_types_of_interest,
  alpha,
  minimum_cells_in_cluster,
  feature_colname = "Cell.Type",
  plot_image = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{cell_types_of_interest}{A character vector specifying the cell types of
interest.}

\item{alpha}{A positive numeric. A smaller alpha value results in clusters
with more intricate borders. A large alpha value results in clusters with
smooth and simple boundaries.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type"}

\item{plot_image}{A logical indicating whether to plot 3D spatial data with
alpha hull clusters. Defaults to TRUE.}
}
\value{
The same 3D SpatialExperiment object used as input for spe, with an
added column in the \code{colData} slot to specify which alpha hull cluster
each cell belongs to, and added metadata containing information needed to
plot the alpha hull clusters.
}
\description{
This function finds cell clusters in a 3D SpatialExperiment
object using the alpha hull clustering algorithm.
}
\examples{
alpha_hull_spe <- alpha_hull_clustering3D(
    spe = SPIAT3D::simulated_spe,
    cell_types_of_interest = c("Tumour", "Immune"),
    alpha = 8,
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

}
