% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_all_single_radius_cc_metrics3D.R
\name{calculate_all_single_radius_cc_metrics3D}
\alias{calculate_all_single_radius_cc_metrics3D}
\title{Apply all cell colocalization metrics on 3D spatial data.}
\usage{
calculate_all_single_radius_cc_metrics3D(
  spe,
  reference_cell_type,
  target_cell_types,
  radius,
  feature_colname
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{reference_cell_type}{A string specifying the reference cell type.}

\item{target_cell_types}{A character vector specifying the target cell types.}

\item{radius}{A positive numeric specifying the radius value.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type".}
}
\value{
A list containing the output of each metric, for each applicable
reference-target cell pair.
}
\description{
This function finds the output of all cell colocalization
metrics on a 3D SpatialExperiment Object. Metrics include: mixing score,
normalized mixing score, neighbourhood counts, cells in neighbourhood,
neighbourhood entropy, cross K, cross L, cross G, fast co-occurrence.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

result <- calculate_all_single_radius_cc_metrics3D(
    spe = simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radius = 30,
    feature_colname = "Cell.Type"
)

}
