% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_border_of_clusters3D.R
\name{calculate_border_of_clusters3D}
\alias{calculate_border_of_clusters3D}
\title{Algorithm to determine cells which form the border of cell clusters in
3D spatial data.}
\usage{
calculate_border_of_clusters3D(
  spe,
  radius,
  cluster_colname,
  feature_colname = "Cell.Type",
  plot_image = T
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell. It must also contain the cell clustering
information, obtained by passing the SpatialExperiment object through one
of the cell clustering algorithm functions in SPIAT-3D
(alpha_hull_clustering3D, grid_based_clustering3D, dbscan_clustering3D).}

\item{cluster_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
clustering information. Should be 'alpha_hull_cluster', 'dbscan_cluster',
or 'grid_based_cluster'.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type"}

\item{plot_image}{A logical indicating whether to plot cell clusters with
bordering cells. Defaults to TRUE.}
}
\value{
The same 3D SpatialExperiment object used as input for spe, with an
added column in the \code{colData} slot called 'cluster_border' to specify
which cells form the border of the clusters, and which cells are outside
or inside the cluster.
}
\description{
This function finds which cells form the border of cell clusters
on in a 3D SpatialExperiment Object, where the cell clusters have already
been identified using an existing SPIAT-3D cell clustering algorithm
function.
}
\examples{
alpha_hull_spe <- alpha_hull_clustering3D(
    spe = SPIAT-3D::simulated_spe,
    cell_types_of_interest = c("Tumour", "Immune"),
    alpha = 8,
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

spe_with_border_of_clusters <- calculate_border_of_clusters3D(
    spe = alpha_hull_spe,
    cluster_colname = "alpha_hull_cluster",
    feature_colname = "Cell.Type",
    plot_image = T
)

}
