% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cell_proportion_grid_metrics3D.R
\name{calculate_cell_proportion_grid_metrics3D}
\alias{calculate_cell_proportion_grid_metrics3D}
\title{Calculate cell proportion grid metrics in 3D spatial data.}
\usage{
calculate_cell_proportion_grid_metrics3D(
  spe,
  n_splits,
  reference_cell_types,
  target_cell_types,
  feature_colname = "Cell.Type",
  plot_image = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{n_splits}{A positive numeric integer specifying the number splits used
to divide the x-axis, y-axis and z-axis.}

\item{reference_cell_types}{A character vector specifying the reference cell
types.}

\item{target_cell_types}{A character vector specifying the target cell types.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type".}

\item{plot_image}{A logical indicating whether to plot cell proportion grid
metrics. Defaults to TRUE.}
}
\value{
A data frame containing the cell proportion and spatial information
for each rectangular prism.
}
\description{
This functions divides a 3D SpatialExperiment Object into a 3D
grid of rectangular prisms. The cell proportions in each rectangular
prism is calculated. The output can be used for other SPIAT-3D metrics:
calculate_spatial_autocorrelation3D, calculate_prevalence3D,
calculate_prevalence_gradient3D.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

cell_prop_grid_metrics <- calculate_cell_proportion_grid_metrics3D(
    spe = simulated_spe,
    n_splits = 10,
    reference_cell_types = c("Tumour"),
    target_cell_types = c("Immune"),
    feature_colname = "Cell.Type",
    plot_image = T
)

}
