% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cell_proportions3D.R
\name{calculate_cell_proportions3D}
\alias{calculate_cell_proportions3D}
\title{Calculate cell proportions in 3D spatial data.}
\usage{
calculate_cell_proportions3D(
  spe,
  cell_types_of_interest = NULL,
  feature_colname = "Cell.Type",
  plot_image = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells.}

\item{cell_types_of_interest}{A character vector specifying the cell types of
interest. If NULL, all cell types in the \code{feature_colname} column will be
considered.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information.}

\item{plot_image}{A logical indicating whether to plot violin plots of the
minimum distances between cell type pairs. Defaults to TRUE.}
}
\value{
A data frame containing the cell types, their frequencies,
proportions, and percentages.
}
\description{
This function calculates the proportions of different cell types
in a 3D SpatialExperiment object. It can optionally plot a bar chart of
the cell proportions.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

cell_proportions <- calculate_cell_proportions3D(
    spe = simulated_spe,
    cell_types_of_interest = NULL,
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

}
