% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cell_proportions_of_clusters3D.R
\name{calculate_cell_proportions_of_clusters3D}
\alias{calculate_cell_proportions_of_clusters3D}
\title{Calculate cell proportion of clusters in 3D spatial data.}
\usage{
calculate_cell_proportions_of_clusters3D(
  spe,
  cluster_colname,
  feature_colname = "Cell.Type",
  plot_image = T
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell. It must also contain the cell clustering
information, obtained by passing the SpatialExperiment object through one
of the cell clustering algorithm functions in SPIAT-3D
(alpha_hull_clustering3D, grid_based_clustering3D, dbscan_clustering3D)}

\item{cluster_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
clustering information. Should be 'alpha_hull_cluster', 'dbscan_cluster',
or 'grid_based_cluster'.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type".}

\item{plot_image}{A logical indicating whether to plot cell proportion of
the cell clusters in bar plot form. Defaults to TRUE.}
}
\value{
A data frame containing the cell proportion of all cell types in
each cell cluster.
}
\description{
This function finds the proportion of cells which make up cell
clusters in a 3D SpatialExperiment Object, where the cell clusters have
already been identified using an existing SPIAT-3D cell clustering
algorithm function.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D")

# Get dbscan clusters
dbscan_spe <- dbscan_clustering3D(
    spe = SPIAT-3D::simulated_spe,
    cell_types_of_interest = c("Tumour", "Immune"),
    radius = 30,
    minimum_cells_in_radius = 10,
    minimum_cells_in_cluster = 30,
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

# Get cell proportions that make up dbscan clusters
cluster_cell_props <- calculate_cell_proportions_of_clusters3D(
    spe = dbscan_spe,
    cluster_colname = "dbscan_cluster",
    feature_colname = "Cell.Type",
    plot_image = T
)

}
