% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cross_G3D.R
\name{calculate_cross_G3D}
\alias{calculate_cross_G3D}
\title{Calculate cross G on 3D spatial data.}
\usage{
calculate_cross_G3D(
  spe,
  reference_cell_type,
  target_cell_type,
  radius,
  feature_colname
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{reference_cell_type}{A string specifying the reference cell type.}

\item{target_cell_type}{A string specifying the target cell type.}

\item{radius}{A positive numeric specifying the radius value.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information.}
}
\value{
A data frame containing observed and expected cross G value.
}
\description{
This function calculates the cross G on a 3D SpatialExperiment
Object. This metric finds the proportion of reference cells which have at
least one interaction with a target cell as well as the expected cross G
values if the 3D spatial data followed a complete spatial randomness
pattern. This is calculated for a single radius value.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

result <- calculate_cross_G3D(
    spe = simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_type = "Immune",
    radius = 5,
    feature_colname = "Cell.Type"
)

}
