% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cross_K_gradient3D.R
\name{calculate_cross_K_gradient3D}
\alias{calculate_cross_K_gradient3D}
\title{Calculate cross K gradient on 3D spatial data.}
\usage{
calculate_cross_K_gradient3D(
  spe,
  reference_cell_type,
  target_cell_types,
  radii,
  feature_colname = "Cell.Type",
  plot_image = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{reference_cell_type}{A string specifying the reference cell type.}

\item{target_cell_types}{A character vector specifying the target cell types.}

\item{radii}{A positive, ascending numeric vector specifying the set of
radius values used.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type".}

\item{plot_image}{A logical indicating whether to plot cross K gradient as a
line graph showing both the observed cross K values for each cell type
and expected cross K values. Defaults to TRUE.}
}
\value{
A data frame containing observed cross K values for each target cell
type and expected cross K values (columns), all across each radii (rows).
}
\description{
This function calculates the cross K gradient on a 3D
SpatialExperiment Object. See paper on theory behind cross K gradient (I
ain't explaining it here).
}
\examples{
result <- calculate_cross_K_gradient3D(
    spe = SPIAT-3D::simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radii = seq(20, 100, 10),
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

}
