% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_entropy_background3D.R
\name{calculate_entropy_background3D}
\alias{calculate_entropy_background3D}
\title{Calculate the background entropy of 3D spatial data.}
\usage{
calculate_entropy_background3D(
  spe,
  cell_types_of_interest,
  feature_colname = "Cell.Type"
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{cell_types_of_interest}{A character vector specifying the cell types of
interest. If NULL, all cell types in the \code{feature_colname} column will be
considered.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type".}
}
\value{
A numeric representing the background entropy.
}
\description{
This function calculates the background entropy of a 3D
SpatialExperiment object across a subset of cell types chosen by the
user.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D")

background_entropy <- calculate_entropy_background3D(
    spe = simulated_spe,
    cell_types_of_interest = c("Tumour", "Immune"),
    feature_colname = "Cell.Type"
)

}
