% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_entropy_grid_metrics3D.R
\name{calculate_entropy_grid_metrics3D}
\alias{calculate_entropy_grid_metrics3D}
\title{Calculate entropy grid metrics in 3D spatial data.}
\usage{
calculate_entropy_grid_metrics3D(
  spe,
  n_splits,
  cell_types_of_interest,
  feature_colname,
  plot_image = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{n_splits}{A positive numeric integer specifying the number splits used
to divide the x-axis, y-axis and z-axis.}

\item{cell_types_of_interest}{A character vector specifying the cell types of
interest.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information.}

\item{plot_image}{A logical indicating whether to plot entropy grid metrics.
Defaults to TRUE.}
}
\value{
A data frame containing the entropy and spatial information for each
rectangular prism.
}
\description{
This functions divides a 3D SpatialExperiment Object into a 3D
grid of rectangular prisms. The entropy values in each rectangular prism
is calculated. The output can be used for other SPIAT3D metrics:
calculate_spatial_autocorrelation3D, calculate_prevalence3D,
calculate_prevalence_gradient3D.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

entropy_grid_metrics <- calculate_entropy_grid_metrics3D(
    spe = simulated_spe,
    n_splits = 10,
    cell_types_of_interest = c("Tumour", "Immune"),
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

}
