% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fast_co_occurrence3D.R
\name{calculate_fast_co_occurrence3D}
\alias{calculate_fast_co_occurrence3D}
\title{Calculate fast co-occurrence on 3D spatial data.}
\usage{
calculate_fast_co_occurrence3D(
  spe,
  reference_cell_type,
  target_cell_types,
  radius,
  feature_colname
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{reference_cell_type}{A string specifying the reference cell type.}

\item{target_cell_types}{A character vector specifying the target cell types.}

\item{radius}{A positive numeric specifying the radius value.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information.}
}
\value{
A data frame containing the fast co-occurrence values for each target
cell type (columns).
}
\description{
This function calculates the fast co-occurrence on a 3D
SpatialExperiment Object. This metric finds the average proportion of
target cells around the reference cells relative to the proportion of
target cells in the SpatialExperiment Object, for each target cell type
and for a single radius value.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

result <- calculate_fast_co_occurrence3D(
    spe = simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radius = 30,
    feature_colname = "Cell.Type"
)

}
