% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_minimum_distances_between_cell_types3D.R
\name{calculate_minimum_distances_between_cell_types3D}
\alias{calculate_minimum_distances_between_cell_types3D}
\title{Calculate minimum distances between cell types in 3D spatial data.}
\usage{
calculate_minimum_distances_between_cell_types3D(
  spe,
  cell_types_of_interest = NULL,
  feature_colname,
  show_summary = TRUE,
  plot_image = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{cell_types_of_interest}{A character vector specifying the cell types of
interest. If NULL, all cell types in the \code{feature_colname} column will be
considered. Defaults to NULL.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information.}

\item{show_summary}{A logical indicating whether to print a summary of the
minimum distances for each cell type pair. Defaults to TRUE.}

\item{plot_image}{A logical indicating whether to plot violin plots of the
minimum distances between cell type pairs. Defaults to TRUE.}
}
\value{
A data frame containing information about the reference cell, the
nearest cell of another type, and the distance between them for each cell
type pair.
}
\description{
This function calculates the minimum distances between different
cell types in a 3D SpatialExperiment object. It allows you to specify a
subset of cell types to analyse and provides the option to summarise the
results and plot violin plots of the minimum distances between cell
types.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

minimum_distances <- calculate_minimum_distances_between_cell_types3D(
    spe = simulated_spe,
    cell_types_of_interest = NULL,
    feature_colname = "Cell.Type",
    show_summary = TRUE,
    plot_image = TRUE
)

}
