% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_minimum_distances_to_clusters3D.R
\name{calculate_minimum_distances_to_clusters3D}
\alias{calculate_minimum_distances_to_clusters3D}
\title{Calculate minimum distances to cell clusters in 3D spatial data.}
\usage{
calculate_minimum_distances_to_clusters3D(
  spe,
  cell_types_inside_cluster,
  cell_types_outside_cluster,
  cluster_colname,
  feature_colname = "Cell.Type",
  plot_image = T
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{cell_types_inside_cluster}{A character vector specifying the cell types
inside the cell clusters that are of interest.}

\item{cell_types_outside_cluster}{A character vector specifying the cell
types outside the cell clusters that are of interest.}

\item{cluster_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
clustering information. Should be 'alpha_hull_cluster', 'dbscan_cluster',
or 'grid_based_cluster'.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type"}

\item{plot_image}{A logical indicating whether to plot violin plots of the
minimum distances each non-cluster cell to each cluster. Defaults to
TRUE.}
}
\value{
A data frame containing information about the reference cell, the
nearest cell of another type, and the distance between them for each cell
type pair.
}
\description{
This function calculates the minimum distances between
non-cluster cells to cluster cells in a  3D SpatialExperiment object,
where the cell clusters have already been identified using an existing
SPIAT-3D cell clustering algorithm function.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

# Get dbscan clusters
dbscan_spe <- dbscan_clustering3D(
    spe = simulated_spe,
    cell_types_of_interest = c("Tumour", "Immune"),
    radius = 8,
    minimum_cells_in_radius = 10,
    minimum_cells_in_cluster = 30,
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

# Get minimum distances to dbscan clusters
minimum_distances_to_clusters <- calculate_minimum_distances_to_clusters3D(
    spe = dbscan_spe,
    cell_types_inside_cluster = c("Tumour"),
    cell_types_outside_cluster = c("Tumour", "Immune"),
    cluster_colname = "dbscan_cluster",
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

}
