% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mixing_scores3D.R
\name{calculate_mixing_scores3D}
\alias{calculate_mixing_scores3D}
\title{Calculate mixing scores on 3D spatial data.}
\usage{
calculate_mixing_scores3D(
  spe,
  reference_cell_types,
  target_cell_types,
  radius,
  feature_colname = "Cell.Type"
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{reference_cell_types}{A character vector specifying the reference cell
types.}

\item{target_cell_types}{A character vector specifying the target cell types.}

\item{radius}{A positive numeric specifying the radius value.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type".}
}
\value{
A data frame containing the mixing score values and associated
information for each valid reference and target cell type combination.
}
\description{
This function calculates the mixing scores on a 3D
SpatialExperiment Object. See paper on theory behind mixing scores (I
ain't explaining it here).
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

result <- calculate_mixing_scores3D(
    spe = simulated_spe,
    reference_cell_types = c("Tumour", "Immune"),
    target_cell_types = c("Tumour", "Immune"),
    radius = 30,
    feature_colname = "Cell.Type"
)

}
