% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_neighbourhood_entropy3D.R
\name{calculate_neighbourhood_entropy3D}
\alias{calculate_neighbourhood_entropy3D}
\title{Calculate neighbourhood entropy gradient on 3D spatial data.}
\usage{
calculate_neighbourhood_entropy3D(
  spe,
  reference_cell_type,
  target_cell_types,
  radius,
  feature_colname = "Cell.Type"
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{reference_cell_type}{A string specifying the reference cell type.}

\item{target_cell_types}{A character vector specifying the target cell types.}

\item{radius}{A positive numeric specifying the radius value.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type".}
}
\value{
A data frame containing the neighbourhood entropy values for each
reference cell (rows) and for each target cell type (columns).
}
\description{
This function calculates the neighbourhood entropy on a 3D
SpatialExperiment Object. This metric finds the entropy between
reference and target cell around each reference cell, for each target
cell type.
}
\examples{
result <- calculate_neighbourhood_entropy3D(
    spe = SPIAT-3D::simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radius = 30,
    feature_colname = "Cell.Type"
)

}
