% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_neighbourhood_entropy_gradient3D.R
\name{calculate_neighbourhood_entropy_gradient3D}
\alias{calculate_neighbourhood_entropy_gradient3D}
\title{Calculate neighbourhood entropy gradient on 3D spatial data.}
\usage{
calculate_neighbourhood_entropy_gradient3D(
  spe,
  reference_cell_type,
  target_cell_types,
  radii,
  feature_colname = "Cell.Type",
  plot_image = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{reference_cell_type}{A string specifying the reference cell type.}

\item{target_cell_types}{A character vector specifying the target cell types.}

\item{radii}{A positive, ascending numeric vector specifying the set of
radius values used.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type".}

\item{plot_image}{A logical indicating whether to plot neighbourhood entropy
gradient as a line graph. Defaults to TRUE.}
}
\value{
A data frame containing the neighbourhood entropy values for each
target cell type (columns) across each radii (rows).
}
\description{
This function calculates the neighbourhood entropy gradient on a
3D SpatialExperiment Object. This metric finds the entropy between
reference and target cell around each reference cell, for each target
cell type, and calculates the average entropy across all radii.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D")

result <- calculate_neighbourhood_entropy_gradient3D(
    spe = simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radii = seq(20, 100, 10),
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

}
