% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prevalence3D.R
\name{calculate_prevalence3D}
\alias{calculate_prevalence3D}
\title{Calculate prevalence of 3D grid metrics.}
\usage{
calculate_prevalence3D(grid_metrics, metric_colname, threshold, above = TRUE)
}
\arguments{
\item{grid_metrics}{A data frame containing the proportion/entropy and
spatial information for each rectangular prism. Obtained from the output
of the calculate_cell_proportion_grid_metrics3D and
calculate_entropy_grid_metrics3D functions.}

\item{metric_colname}{A string specifying the name of the column in
\code{grid_metrics} containing the proportion/entropy information. Should be
'proportion' or 'entropy'.}

\item{threshold}{A numeric between 0 and 1 representing the threshold value.}

\item{above}{A logical indicating whether to calculate the proportion of
rectangular prisms which surpass the threshold (TRUE) or do not (FALSE).
Defaults to TRUE.}
}
\value{
A numeric representing the prevalence value.
}
\description{
This functions calculates the 'prevalence', or equivalently,
the proportion of rectangular prisms of the 3D grid metrics has a value
above/below the required threshold.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D")

# Get cell proportion grid metrics
cell_prop_grid_metrics <- calculate_cell_proportion_grid_metrics3D(
    spe = simulated_spe,
    n_splits = 10,
    reference_cell_types = c("Tumour"),
    target_cell_types = c("Immune"),
    feature_colname = "Cell.Type",
    plot_image = T
)

# Get prevalence from cell proportion grid metrics
prevalence <- calculate_prevalence3D(
    grid_metrics = cell_prop_grid_metrics,
    metric_colname = "proportion",
    threshold = 0.3,
    above = TRUE
)

}
