% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prevalence_gradient3D.R
\name{calculate_prevalence_gradient3D}
\alias{calculate_prevalence_gradient3D}
\title{Calculate prevalence gradient of 3D grid metrics.}
\usage{
calculate_prevalence_gradient3D(
  grid_metrics,
  metric_colname,
  show_AUC = T,
  plot_image = T
)
}
\arguments{
\item{grid_metrics}{A data frame containing the proportion/entropy and
spatial information for each rectangular prism. Obtained from the output
of the calculate_cell_proportion_grid_metrics3D and
calculate_entropy_grid_metrics3D functions.}

\item{metric_colname}{A string specifying the name of the column in
\code{grid_metrics} containing the proportion/entropy information. Should be
'proportion' or 'entropy'.}

\item{show_AUC}{A logical indicating whether to print out the prevalence
gradient AUC value. Defaults to TRUE.}

\item{plot_image}{A logical indicating whether to plot a line graph showing
prevalence vs threshold values. Defaults to TRUE.}
}
\value{
A data frame containing the prevalence values for each threshold.
}
\description{
This functions calculates the 'prevalence gradient', or
equivalently, the proportion of rectangular prisms of the 3D grid metrics
has a value above the required threshold, for a gradient of thresholds.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D")

# Get cell proportion grid metrics
cell_prop_grid_metrics <- calculate_cell_proportion_grid_metrics3D(
    spe = simulated_spe,
    n_splits = 10,
    reference_cell_types = c("Tumour"),
    target_cell_types = c("Immune"),
    feature_colname = "Cell.Type",
    plot_image = T
)

# Get prevalence gradient from cell proportion grid metrics
prevalence_gradient <- calculate_prevalence_gradient3D(
    grid_metrics = cell_prop_grid_metrics,
    metric_colname = "proportion",
    show_AUC = T,
    plot_image = TRUE
)

}
