% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan_clustering3D.R
\name{dbscan_clustering3D}
\alias{dbscan_clustering3D}
\title{Find cell clusters in 3D spatial data using dbscan clustering
algorithm.}
\usage{
dbscan_clustering3D(
  spe,
  cell_types_of_interest,
  radius,
  minimum_cells_in_radius,
  minimum_cells_in_cluster,
  feature_colname = "Cell.Type",
  plot_image = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{cell_types_of_interest}{A character vector specifying the cell types of
interest.}

\item{radius}{A positive numeric. Spheres of specified radius are drawn
around each cell type of interest and the number of other cell types of
interest are counted.}

\item{minimum_cells_in_radius}{A positive numeric. If the number of cells
types of interest within the sphere of another cell type of interest
surpasses this specified value, they form a cluster.}

\item{minimum_cells_in_cluster}{A positive numeric. Clusters identified with
dbscan which have less than this specified value are relabelled as not a
cluster.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type"}

\item{plot_image}{A logical indicating whether to plot 3D spatial data with
alpha hull clusters. Defaults to TRUE.}
}
\value{
The same 3D SpatialExperiment object used as input for spe, with an
added column in the \code{colData} slot to specify which dbscan cluster each
cell belongs to.
}
\description{
This function finds cell clusters in a 3D SpatialExperiment
object using the dbscan clustering algorithm.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

dbscan_spe <- dbscan_clustering3D(
    spe = simulated_spe,
    cell_types_of_interest = c("Tumour", "Immune"),
    radius = 30,
    minimum_cells_in_radius = 10,
    minimum_cells_in_cluster = 30,
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

}
