% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spe_grid_metrics3D.R
\name{get_spe_grid_metrics3D}
\alias{get_spe_grid_metrics3D}
\title{Calculate grid metrics in 3D spatial data.}
\usage{
get_spe_grid_metrics3D(spe, n_splits, feature_colname)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{n_splits}{A positive numeric integer specifying the number of splits
used to divide the x-axis, y-axis and z-axis.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information.}
}
\value{
A list containing the number and type of cells in each rectangular
prism, and the coordinate of each rectangular prism.
}
\description{
This functions divides a 3D SpatialExperiment Object into a 3D
grid of rectangular prisms. The number and types of cells in each
rectangular prism and the coordinate of each rectangular prism is
calculated. The is a supporting function used in
calculate_cell_proportion_grid_metrics3D,
calculate_entropy_grid_metrics3D and grid_based_clustering3D.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

spe_with_grid_metrics <- get_spe_grid_metrics3D(
    spe = simulated_spe,
    n_splits = 10,
    feature_colname = "Cell.Type"
)

}
