% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_based_clustering3D.R
\name{grid_based_clustering3D}
\alias{grid_based_clustering3D}
\title{Find cell clusters in 3D spatial data using grid based clustering
algorithm.}
\usage{
grid_based_clustering3D(
  spe,
  cell_types_of_interest,
  n_splits,
  minimum_cells_in_cluster,
  feature_colname = "Cell.Type",
  plot_image = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{cell_types_of_interest}{A character vector specifying the cell types of
interest.}

\item{n_splits}{A positive numeric integer specifying the number splits used
to divide the x-axis, y-axis and z-axis.}

\item{minimum_cells_in_cluster}{A positive numeric. Clusters identified which
have less than this specified value are relabelled as not a cluster.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type"}

\item{plot_image}{A logical indicating whether to plot 3D spatial data with
grid based clusters. Defaults to TRUE.}
}
\value{
The same 3D SpatialExperiment object used as input for spe, with an
added column in the \code{colData} slot to specify which grid based cluster
each cell belongs to, and added metadata containing information needed to
plot the grid based clusters.
}
\description{
This function finds cell clusters in a 3D SpatialExperiment
object using the grid based clustering algorithm.
}
\examples{
grid_based_spe <- grid_based_clustering3D(
    spe = SPIAT-3D::simulated_spe,
    cell_types_of_interest = c("Tumour", "Immune"),
    n_splits = 10,
    minimum_cells_in_cluster = 30,
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

}
