% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_co_occurrence_gradient3D.R
\name{plot_co_occurrence_gradient3D}
\alias{plot_co_occurrence_gradient3D}
\title{Function to plot co-occurrence gradient data.}
\usage{
plot_co_occurrence_gradient3D(co_occurrence_gradient_df)
}
\arguments{
\item{co_occurrence_gradient_df}{A data frame obtained from the output of the
calculate_co_occurrence_gradient3D function.}
}
\value{
A ggplot object containing the line graph.
}
\description{
This function plots the co-occurrence gradient data as a line
graph, showing co-occurrence vs radius.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

# Calculate co-occurrence gradient from simulated spe
result <- calculate_co_occurrence_gradient3D(
    spe = simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radii = seq(20, 100, 10),
    feature_colname = "Cell.Type",
    plot_image = FALSE
)

# Plot
fig <- plot_co_occurrence_gradient3D(
    co_occurrence_gradient_df = result
)

methods::show(fig)

}
