% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cross_G_gradient3D.R
\name{plot_cross_G_gradient3D}
\alias{plot_cross_G_gradient3D}
\title{Function to plot cross G gradient data.}
\usage{
plot_cross_G_gradient3D(
  cross_G_gradient_df,
  reference_cell_type = NULL,
  target_cell_type = NULL
)
}
\arguments{
\item{cross_G_gradient_df}{A data frame obtained from the output of the
calculate_cross_G_gradient3D function.}

\item{reference_cell_type}{A string specifying the reference cell type. If
NULL or if \code{target_cell_type} is NULL, no reference cell type label will
be showed on the plot. Defaults to NULL.}

\item{target_cell_type}{A string specifying the target cell type. If NULL or
if \code{reference_cell_type} is NULL, no target cell type label will be
showed on the plot. Defaults to NULL.}
}
\value{
A ggplot object containing the line graph.
}
\description{
This function plots the cross G gradient data as a line graph,
showing cross G vs radius.
}
\examples{
result <- calculate_cross_G_gradient3D(
    spe = SPIAT-3D::simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_type = "Immune",
    radii = seq(20, 100, 10),
    feature_colname = "Cell.Type",
    plot_image = FALSE
)

fig <- plot_cross_G_gradient3D(
    cross_G_gradient_df = result,
    reference_cell_type = "Tumour",
    target_cell_type = "Immune"
)

methods::show(fig)

}
