% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cross_K_gradient3D.R
\name{plot_cross_K_gradient3D}
\alias{plot_cross_K_gradient3D}
\title{Function to plot cross K gradient data.}
\usage{
plot_cross_K_gradient3D(cross_K_gradient_df)
}
\arguments{
\item{cross_K_gradient_df}{A data frame obtained from the output of the
calculate_cross_K_gradient3D function.}
}
\value{
A ggplot object containing the line graph.
}
\description{
This function plots the cross K gradient data as a line graph,
showing cross K vs radius.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

# Calculate cross K gradient from simulated spe
result <- calculate_cross_K_gradient3D(
    spe = simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radii = seq(20, 100, 10),
    feature_colname = "Cell.Type",
    plot_image = FALSE
)

# Plot
fig <- plot_cross_K_gradient3D(
    cross_K_gradient_df = result
)

methods::show(fig)

}
