% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cross_L_gradient3D.R
\name{plot_cross_L_gradient3D}
\alias{plot_cross_L_gradient3D}
\title{Function to plot cross L gradient data.}
\usage{
plot_cross_L_gradient3D(cross_L_gradient_df)
}
\arguments{
\item{cross_L_gradient_df}{A data frame obtained from the output of the
calculate_cross_L_gradient3D function.}
}
\value{
A ggplot object containing the line graph.
}
\description{
This function plots the cross L gradient data as a line graph,
showing cross L vs radius.
}
\examples{
result <- calculate_cross_L_gradient3D(
    spe = SPIAT-3D::simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radii = seq(20, 100, 10),
    feature_colname = "Cell.Type",
    plot_image = FALSE
)

fig <- plot_cross_L_gradient3D(
    cross_L_gradient_df = result
)

methods::show(fig)

}
