% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distances_between_cell_types_violin3D.R
\name{plot_distances_between_cell_types_violin3D}
\alias{plot_distances_between_cell_types_violin3D}
\title{Function to plot distances between cell types as a set of violin
plots.}
\usage{
plot_distances_between_cell_types_violin3D(distances_df, scales = "free_x")
}
\arguments{
\item{distances_df}{A data frame obtained from the output of the
calculate_pairwise_distances_between_cell_types3D or
calculate_minimum_distances_between_cell_types3D functions.}

\item{scales}{A string to control the facet wrap scales of the ggplot object.
Either "fixed", "free_x", "free_y" or "free". Would not recommend
"free_y". Defaults to "free_x".}
}
\value{
A ggplot object containing the violin plots.
}
\description{
This function plots the distances between cell types data as
violin plots, showing the distances for each cell type pair.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

# Calculate minimum distances from simulated spe
minimum_distances <- calculate_minimum_distances_between_cell_types3D(
    spe = simulated_spe,
    cell_types_of_interest = NULL,
    feature_colname = "Cell.Type",
    show_summary = TRUE,
    plot_image = FALSE
)

# Plot
fig <- plot_distances_between_cell_types_violin3D(
    distances_df = minimum_distances,
    scales = "free_x"
)

methods::show(fig)

}
