% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid_based_clusters3D.R
\name{plot_grid_based_clusters3D}
\alias{plot_grid_based_clusters3D}
\title{Function to plot grid based clusters in 3D spatial data.}
\usage{
plot_grid_based_clusters3D(
  spe_with_grid,
  plot_cell_types = NULL,
  plot_colours = NULL,
  feature_colname = "Cell.Type"
)
}
\arguments{
\item{plot_cell_types}{A string vector specifying the cell types to plot. If
NULL, all cell types in the \code{feature_colname} column will be considered.
Defaults to NULL.}

\item{plot_colours}{A string vector specifying the colours of the cell types
when plotting. Must match the number of cell types specified in
\code{plot_cell_types}. If NULL, the rainbow color pallete will be used.
Defaults to NULL.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type"}

\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell. It must also contain the cell clustering
information, obtained by passing the SpatialExperiment object through the
grid_based_clustering3D function.}
}
\value{
A Plotly object plotting the cells and grid based clusters of the 3D
SpatialExperiment Object.
}
\description{
This functions plots the cells of 3D spatial data and also
displays the grid based clusters in a 3D SpatialExperiment Object where
grid based clusters have already been identified.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D")

# Get grid based clusters
grid_based_spe <- grid_based_clustering3D(
    spe = simulated_spe,
    cell_types_of_interest = c("Tumour", "Immune"),
    n_splits = 10,
    minimum_cells_in_cluster = 30,
    feature_colname = "Cell.Type",
    plot_image = FALSE
)


# Plot
fig <- plot_grid_based_clusters3D(
    spe_with_grid = grid_based_spe,
    plot_cell_types = NULL,
    plot_colours = NULL,
    feature_colname = "Cell.Type"
)

methods::show(fig)

}
