% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid_metrics_continuous3D.R
\name{plot_grid_metrics_continuous3D}
\alias{plot_grid_metrics_continuous3D}
\title{Plot 3D grid metrics (continuous method).}
\usage{
plot_grid_metrics_continuous3D(grid_metrics, metric_colname)
}
\arguments{
\item{grid_metrics}{A data frame containing the proportion/entropy and
spatial information for each rectangular prism. Obtained from the output
of the calculate_cell_proportion_grid_metrics3D and
calculate_entropy_grid_metrics3D functions.}

\item{metric_colname}{A string specifying the name of the column in
\code{grid_metrics} containing the proportion/entropy information. Should be
'proportion' or 'entropy'.}
}
\value{
A Plotly object containing the 3D grid metrics plot.
}
\description{
This functions plots the 3D grid metrics, showing the
proportion/entropy values of each rectangular prism using a continuous
color scale for the proportion/entropy values.
}
\examples{
cell_prop_grid_metrics <- calculate_cell_proportion_grid_metrics3D(
    spe = SPIAT-3D::simulated_spe,
    n_splits = 10,
    reference_cell_types = c("Tumour"),
    target_cell_types = c("Immune"),
    feature_colname = "Cell.Type",
    plot_image = FALSE
)

fig <- plot_grid_metrics_continuous3D(
    grid_metrics = cell_prop_grid_metrics,
    metric_colname = "proportion"
)

methods::show(fig)

}
