% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neighbourhood_counts_gradient3D.R
\name{plot_neighbourhood_counts_gradient3D}
\alias{plot_neighbourhood_counts_gradient3D}
\title{Function to plot neighbourhood counts gradient data.}
\usage{
plot_neighbourhood_counts_gradient3D(
  neighbourhood_counts_gradient_df,
  reference_cell_type = NULL
)
}
\arguments{
\item{neighbourhood_counts_gradient_df}{A data frame obtained from the output
of calculate_neighbourhood_counts3D function.}

\item{reference_cell_type}{A string specifying the reference cell type. If
NULL, no reference cell type label will be showed on the plot. Defaults
to NULL.}
}
\value{
A ggplot object containing the line graph.
}
\description{
This function plots the neighbourhood counts gradient data as a
line graph, showing average neighbourhood counts vs radius.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

# Calculate neighbourhood counts gradient from simulated spe
result <- calculate_neighbourhood_counts_gradient3D(
    spe = simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radii = seq(20, 100, 10),
    feature_colname = "Cell.Type",
    plot_image = FALSE
)

# Plot
fig <- plot_neighbourhood_counts_gradient3D(
    neighbourhood_counts_gradient_df = result,
    reference_cell_type = "Tumour"
)

methods::show(fig)

}
