% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neighbourhood_counts_violin3D.R
\name{plot_neighbourhood_counts_violin3D}
\alias{plot_neighbourhood_counts_violin3D}
\title{Function to plot neighbourhood counts data as a set of violin plots.}
\usage{
plot_neighbourhood_counts_violin3D(
  neighbourhood_counts_df,
  reference_cell_type,
  scales = "free_x"
)
}
\arguments{
\item{neighbourhood_counts_df}{A data frame obtained from the output of the
calculate_neighbourhood_counts3D function.}

\item{reference_cell_type}{A string specifying the reference cell type.}

\item{scales}{A string to control the facet wrap scales of the ggplot object.
Either "fixed", "free_x", "free_y" or "free". Would not recommend
"free_y". Defaults to "free_x".}
}
\value{
A ggplot object containing the violin plots.
}
\description{
This function plots the neighbourhood counts data as violin
plots, showing the neighbourhood counts for each reference and target
cell type pair.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D")

# Calculate neighbourhood counts from spe
result <- calculate_neighbourhood_counts3D(
    spe = simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radius = 30,
    feature_colname = "Cell.Type",
    show_summary = TRUE,
    plot_image = FALSE
)

# Plot
fig <- plot_neighbourhood_counts_violin3D(
    neighbourhood_counts_df = result,
    reference_cell_type = "Tumour",
    scales = "free_x"
)

methods::show(fig)

}
