% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_distances_between_cell_types3D.R
\name{summarise_distances_between_cell_types3D}
\alias{summarise_distances_between_cell_types3D}
\title{Summarise distances between cell types.}
\usage{
summarise_distances_between_cell_types3D(distances_df)
}
\arguments{
\item{distances_df}{A data frame obtained from the output of the
calculate_pairwise_distances_between_cell_types3D or
calculate_minimum_distances_between_cell_types3D functions.}
}
\value{
A data frame containing the summarised data.
}
\description{
This function summarises the distances between cell types data,
showing mean, min, max, median and standard deviation for each cell type
pair.
}
\examples{
minimum_distances <- calculate_minimum_distances_between_cell_types3D(
    spe = SPIAT-3D::simulated_spe,
    cell_types_of_interest = NULL,
    feature_colname = "Cell.Type",
    show_summary = FALSE,
    plot_image = FALSE
)

distances_summary <- summarise_distances_between_cell_types3D(
    distances_df = minimum_distances
)

}
