% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_neighbourhood_metric3D.R
\name{summarise_neighbourhood_metric3D}
\alias{summarise_neighbourhood_metric3D}
\title{Summarise neighbourhood metric data.}
\usage{
summarise_neighbourhood_metric3D(neighbourhood_metric_df)
}
\arguments{
\item{neighbourhood_metric_df}{A data frame obtained from the output of
either the calculate_neighbourhood_counts3D,
calculate_cells_in_neighbourhood3D or calculate_neighbourhood_entropy3D
function.}
}
\value{
A data frame containing the summarised data.
}
\description{
This function summarises the data from either neighbourhood
counts, cells in neighbourhood or neighbourhood entropy, showing, min,
max, median and standard deviation for each reference and target cell
type pair.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))

# Calculate neighbourhood counts from simulated spe
result <- calculate_neighbourhood_counts3D(
    spe = simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radius = 30,
    feature_colname = "Cell.Type"
)

# Get summary
neighbourhood_counts_summary <- summarise_neighbourhood_metric3D(
    neighbourhood_metric_df = result
)

}
