% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_all_gradient_cc_metrics3D.R
\name{calculate_all_gradient_cc_metrics3D}
\alias{calculate_all_gradient_cc_metrics3D}
\title{Apply all cell colocalization metrics (gradient version) on 3D spatial
data.}
\usage{
calculate_all_gradient_cc_metrics3D(
  spe,
  reference_cell_type,
  target_cell_types,
  radii,
  feature_colname = "Cell.Type",
  plot_image = T
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{reference_cell_type}{A string specifying the reference cell type.}

\item{target_cell_types}{A character vector specifying the target cell types.}

\item{radii}{A positive, ascending numeric vector specifying the set of
radius values used to calculate each metric over a gradient}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type"}

\item{plot_image}{A logical indicating whether to plot analysis of all
metrics. Defaults to TRUE.}
}
\value{
A list containing the output of each metric, for each applicable
reference-target cell pair.
}
\description{
This function finds the output of all cell colocalization
metrics (gradient version) on a 3D SpatialExperiment Object. Metrics
include: mixing score, normalized mixing score, neighbourhood counts,
cells in neighbourhood, neighbourhood entropy, cross K, cross L, cross G,
co-occurrence.
}
\examples{
result <- calculate_all_gradient_cc_metrics3D(
    spe = SPIAT-3D::simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radii = seq(20, 100, 10),
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

}
