% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_co_occurrence_gradient3D.R
\name{calculate_co_occurrence_gradient3D}
\alias{calculate_co_occurrence_gradient3D}
\title{Calculate co-occurrence gradient on 3D spatial data.}
\usage{
calculate_co_occurrence_gradient3D(
  spe,
  reference_cell_type,
  target_cell_types,
  radii,
  feature_colname = "Cell.Type",
  plot_image = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{reference_cell_type}{A string specifying the reference cell type.}

\item{target_cell_types}{A character vector specifying the target cell types.}

\item{radii}{A positive, ascending numeric vector specifying the set of
radius values used to calculate cells in neighbourhood over a gradient.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type".}

\item{plot_image}{A logical indicating whether to plot co-occurrence gradient
as a line graph. Defaults to TRUE.}
}
\value{
A data frame containing the co-occurrence values for each target cell
type (columns) across each radii (rows).
}
\description{
This function calculates the co-occurrence gradient on a 3D
SpatialExperiment Object. This metric finds the proportion of target
cells around reference cells relative to the proportion of target cells
in the SpatialExperiment Object, for each target cell type, and across
all radii values.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D")

result <- calculate_co_occurrence_gradient3D(
    spe = simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_types = c("Tumour", "Immune"),
    radii = seq(20, 100, 10),
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

}
