% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mixing_scores_gradient3D.R
\name{calculate_mixing_scores_gradient3D}
\alias{calculate_mixing_scores_gradient3D}
\title{Calculate mixing scores gradient on 3D spatial data.}
\usage{
calculate_mixing_scores_gradient3D(
  spe,
  reference_cell_type,
  target_cell_type,
  radii,
  feature_colname = "Cell.Type",
  plot_image = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{reference_cell_type}{A string specifying the reference cell type.}

\item{target_cell_type}{A string specifying the target cell type.}

\item{radii}{A positive, ascending numeric vector specifying the set of
radius values used.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type".}

\item{plot_image}{A logical indicating whether to plot mixing scores gradient
as a  line graph. Defaults to TRUE.}
}
\value{
A data frame containing mixing score values and associated
information across each radii (rows).
}
\description{
This function calculates the mixing scores gradient on a 3D
SpatialExperiment Object. See paper on theory behind mixing scores
gradient (I ain't explaining it here).
}
\examples{
result <- calculate_mixing_scores_gradient3D(
    spe = SPIAT-3D::simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_type = "Immune",
    radii = seq(20, 100, 10),
    feature_colname = "Cell.Type",
    plot_image = TRUE
)

}
