% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cells3D.R
\name{plot_cells3D}
\alias{plot_cells3D}
\title{Function to plot cells in 3D spatial data.}
\usage{
plot_cells3D(
  spe,
  plot_cell_types = NULL,
  plot_colours = NULL,
  feature_colname = "Cell.Type"
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing 3D spatial information for
the cells. Naming of spatial coordinates MUST be "Cell.X.Position",
"Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate
and z-coordinate of each cell.}

\item{plot_cell_types}{A string vector specifying the cell types to plot. If
NULL, all cell types in the \code{feature_colname} column will be considered.
Defaults to NULL.}

\item{plot_colours}{A string vector specifying the colours of the cell types
when plotting. Must match the number of cell types specified in
\code{plot_cell_types}. If NULL, the rainbow color pallete will be used.
Defaults to NULL.}

\item{feature_colname}{A string specifying the name of the column in the
\code{colData} slot of the SpatialExperiment object that contains the cell
type information. Defaults to "Cell.Type"}
}
\value{
A Plotly object plotting the cells of the 3D SpatialExperiment
Object.
}
\description{
This functions plots the cells of 3D spatial data in a 3D
SpatialExperiment Object.
}
\examples{
fig <- plot_cells3D(
    spe = SPIAT-3D::simulated_spe,
    plot_cell_types = NULL,
    plot_colours = NULL,
    feature_colname = "Cell.Type"
)

methods::show(fig)

}
