% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixing_scores_gradient3D.R
\name{plot_mixing_scores_gradient3D}
\alias{plot_mixing_scores_gradient3D}
\title{Function to plot mixing scores gradient data.}
\usage{
plot_mixing_scores_gradient3D(mixing_scores_gradient_df, metric = "MS")
}
\arguments{
\item{mixing_scores_gradient_df}{A data frame obtained from the
output of calculate_mixing_scores_gradient3D function.}

\item{metric}{A string specifying the chosen metric. Either "MS" (mixing
score) or "NMS" (normalised mixing score).}
}
\value{
A ggplot object containing the line graph.
}
\description{
This function plots the mixing scores gradient data as
a line graph, showing either mixing scores or normalised mixing scores vs
radius.
}
\examples{
# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D")

# Calculate mixing scores gradient from simulated spe
result <- calculate_mixing_scores_gradient3D(
    spe = simulated_spe,
    reference_cell_type = "Tumour",
    target_cell_type = "Immune",
    radii = seq(20, 100, 10),
    feature_colname = "Cell.Type",
    plot_image = FALSE
)

# Plot
fig <- plot_mixing_scores_gradient3D(
    mixing_scores_gradient_df = result,
    metric = "MS"
)

methods::show(fig)

}
